Move PathFilters from internal to public API
