+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/PathFilters.kt
@@ -1,9 +1,8 @@
-package io.gitlab.arturbosch.detekt.api.internal
+package io.gitlab.arturbosch.detekt.api
 
 import io.github.detekt.psi.absolutePath
 import io.github.detekt.psi.basePath
-import io.gitlab.arturbosch.detekt.api.Config
-import io.gitlab.arturbosch.detekt.api.commaSeparatedPattern
+import io.gitlab.arturbosch.detekt.api.internal.pathMatcher
 import org.jetbrains.kotlin.psi.KtFile
 import java.nio.file.Path
 import java.nio.file.PathMatcher
